refactor(stt): move the Soniox client into its own module
streaming.py was already over the product line-count ratchet before this branch;
adding a provider inline pushed it further. The Soniox protocol is self-contained,
so it lives in utils/stt/soniox.py and streaming.py re-exports it for existing
import sites. Imports flow one way, so there is no cycle.

Also drops a redundant isinstance on a typed dict that pyright rejected, and
annotates JSON-sourced tokens as Any so the remaining runtime guard is real.
